refactor(server): split sessions.rs inline tests into sessions_tests.rs (1,000-line cap)
The inline test module in sessions.rs (755 lines) exceeded the 1,000-line
structural cap. Split tests into a separate module using the #[path]
include pattern (identical to tabs_persist.rs:997-999).

- sessions.rs: 1189 → 437 lines (production code only: ~434 lines)
- sessions_tests.rs: 752 lines (test module content)
- All 25 sessions tests + 9 generate_title tests pass unchanged
- cargo fmt check clean on both files
